Move code used in admin only
