+++ frontend/src/metabase/admin/datamodel/components/QueryDefinitionTooltip/index.ts
@@ -1,0 +1,1 @@
+export * from "./QueryDefinitionTooltip";
