+++ frontend/src/metabase/query_builder/components/AggregationWidget.jsx
@@ -5,8 +5,8 @@ import { t } from "ttag";
 
 import ControlledPopoverWithTrigger from "metabase/components/PopoverWithTrigger/ControlledPopoverWithTrigger";
 import { isRows } from "metabase-lib/queries/utils/aggregation";
+import AggregationPopover from "../../admin/datamodel/components/AggregationPopover";
 import Clearable from "./Clearable";
-import AggregationPopover from "./AggregationPopover";
 import { AggregationLabel } from "./AggregationWidget.styled";
 // NOTE: lots of duplication between AggregationWidget and BreakoutWidget
